updated the footer
