+++ src/Layout/RootLayout.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 import Header from '../components/Header'
 import Footer from '../components/Footer'
 import { Outlet } from 'react-router'
-import TUIFooter from '../components/TUIFooter'
 
 const RootLayout = () => {
   return (
@@ -11,8 +10,7 @@ const RootLayout = () => {
       <main className='bg-white px-6 min-h-screen'>
         <Outlet />
       </main>
-      {/* <Footer /> */}
-      <TUIFooter/>
+      <Footer />
     </>
   )
 }
